test(test262): split dedicated project
Move test262-specific bootstrap/parser coverage into a dedicated Js2IL.Test262.Tests project and extract shared execution/generator test infrastructure into Js2IL.Testing so the main test assembly stays smaller.
