+++ js2il.sln
@@ -13,6 +13,16 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Js2IL.Core", "src\Js2IL.Cor
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Js2IL.SDK", "src\Js2IL.SDK\Js2IL.SDK.csproj", "{25CFA47E-2134-45A5-8935-5A5DBDC8082D}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{0AB3BF05-4346-4AA6-1389-037BE0695223}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Js2IL.Testing", "tests\Js2IL.Testing\Js2IL.Testing.csproj", "{EF25232A-0200-41FD-85BF-B54DE6A6F839}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{827E0CD3-B72D-47B6-A68D-7590B98EB39B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Cli", "Cli", "{778427DF-EB2C-5228-964E-6D535DE253CA}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Js2IL.Test262.Tests", "tests\Js2IL.Test262.Tests\Js2IL.Test262.Tests.csproj", "{0D099E04-F382-4606-9381-AEAE6C355173}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -71,8 +81,37 @@ Global
 		{25CFA47E-2134-45A5-8935-5A5DBDC8082D}.Release|x64.Build.0 = Release|Any CPU
 		{25CFA47E-2134-45A5-8935-5A5DBDC8082D}.Release|x86.ActiveCfg = Release|Any CPU
 		{25CFA47E-2134-45A5-8935-5A5DBDC8082D}.Release|x86.Build.0 = Release|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Debug|x64.Build.0 = Debug|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Debug|x86.Build.0 = Debug|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Release|x64.ActiveCfg = Release|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Release|x64.Build.0 = Release|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Release|x86.ActiveCfg = Release|Any CPU
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839}.Release|x86.Build.0 = Release|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Debug|x64.Build.0 = Debug|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Debug|x86.Build.0 = Debug|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Release|x64.ActiveCfg = Release|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Release|x64.Build.0 = Release|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Release|x86.ActiveCfg = Release|Any CPU
+		{0D099E04-F382-4606-9381-AEAE6C355173}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{EF25232A-0200-41FD-85BF-B54DE6A6F839} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{778427DF-EB2C-5228-964E-6D535DE253CA} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{0D099E04-F382-4606-9381-AEAE6C355173} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 	EndGlobalSection
 EndGlobal
