+++ src/JavaScriptRuntime/JavaScriptRuntime.csproj
@@ -24,6 +24,7 @@
 
   <ItemGroup>
     <InternalsVisibleTo Include="Js2IL.Tests" />
+    <InternalsVisibleTo Include="Js2IL.Testing" />
   </ItemGroup>
 
   <ItemGroup>
