+++ src/Js2IL.Core/Js2IL.Core.csproj
@@ -25,6 +25,7 @@
 
   <ItemGroup>
     <InternalsVisibleTo Include="Js2IL.Tests" />
+    <InternalsVisibleTo Include="Js2IL.Testing" />
   </ItemGroup>
 
   <ItemGroup>
