+++ tests/Js2IL.Test262.Tests/Integration/Snapshots/ExecutionTests.Compile_Scripts_Test262Bootstrap.verified.txt
