+++ tests/Js2IL.Testing/MockFileSystem.cs
