+++ tests/Js2IL.Testing/TestLogger.cs
