+++ tests/Js2IL.Testing/VerifyModuleInitializer.cs
