+++ tests/Js2IL.Tests/Integration/GeneratorTests.cs
@@ -16,13 +16,6 @@ namespace Js2IL.Tests.Integration
         public Task Compile_Scripts_DecompileGeneratorTest() => GenerateTest(nameof(Compile_Scripts_DecompileGeneratorTest));
 
         [Fact]
-        public Task Compile_Scripts_Test262Bootstrap() => GenerateTest(nameof(Compile_Scripts_Test262Bootstrap));
-
-        [Fact]
-        public Task Compile_Scripts_Test262MetadataParser()
-            => GenerateTest(nameof(Compile_Scripts_Test262MetadataParser), ["test262/metadataParser"]);
-
-        [Fact]
         public Task Compile_Scripts_ConvertEcmaExtractHtmlToMarkdown()
             => GenerateTest(nameof(Compile_Scripts_ConvertEcmaExtractHtmlToMarkdown), ["node_modules/turndown/index"]);
