+++ tests/Js2IL.Tests/Js2IL.Tests.csproj
@@ -38,12 +38,6 @@
     <EmbeddedResource Include="..\..\scripts\decompileGeneratorTest.js">
       <LogicalName>Js2IL.Tests.Integration.JavaScript.Compile_Scripts_DecompileGeneratorTest.js</LogicalName>
     </EmbeddedResource>
-    <EmbeddedResource Include="..\..\scripts\test262\bootstrap.js">
-      <LogicalName>Js2IL.Tests.Integration.JavaScript.Compile_Scripts_Test262Bootstrap.js</LogicalName>
-    </EmbeddedResource>
-    <EmbeddedResource Include="..\..\scripts\test262\metadataParser.js">
-      <LogicalName>Js2IL.Tests.Integration.JavaScript.test262.metadataParser.js</LogicalName>
-    </EmbeddedResource>
     <EmbeddedResource Include="Integration\Resources\extractEcma262SectionHtml_urlMode.js">
       <LogicalName>Js2IL.Tests.Integration.JavaScript.Compile_Scripts_ExtractEcma262SectionHtml_UrlMode.js</LogicalName>
     </EmbeddedResource>
@@ -52,9 +46,6 @@
     </EmbeddedResource>
     <EmbeddedResource Include="Integration\Resources\extractEcma262SectionHtml_testHarness.js">
       <LogicalName>Js2IL.Tests.Integration.JavaScript.Compile_Scripts_ExtractEcma262SectionHtml_TestHarness.js</LogicalName>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Integration\Resources\test262MetadataParser_testHarness.js">
-      <LogicalName>Js2IL.Tests.Integration.JavaScript.Compile_Scripts_Test262MetadataParser.js</LogicalName>
     </EmbeddedResource>
     <EmbeddedResource Include="Integration\Resources\turndown_index.js">
       <LogicalName>Js2IL.Tests.Integration.JavaScript.node_modules.turndown.index.js</LogicalName>
@@ -88,6 +79,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Js2IL.Testing\Js2IL.Testing.csproj" />
     <ProjectReference Include="..\..\src\Js2IL.Core\Js2IL.Core.csproj" />
     <ProjectReference Include="..\..\src\Cli\Js2IL.csproj" />
   </ItemGroup>
