+++ tests/Js2IL.Tests/Js2IL.Tests.sln
@@ -1,19 +1,102 @@
+﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Js2IL.Tests", "Js2IL.Tests.csproj", "{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Js2IL.Testing", "..\Js2IL.Testing\Js2IL.Testing.csproj", "{6329159F-A5E2-4E45-B976-545BFE791474}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Js2IL.Core", "..\..\src\Js2IL.Core\Js2IL.Core.csproj", "{1F40F9F4-6402-4F73-982E-F5A388868651}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JavaScriptRuntime", "..\..\src\JavaScriptRuntime\JavaScriptRuntime.csproj", "{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Js2IL", "..\..\src\Cli\Js2IL.csproj", "{CA34D222-9A1E-41EC-BD06-68E146505A6B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Js2IL.Test262.Tests", "..\Js2IL.Test262.Tests\Js2IL.Test262.Tests.csproj", "{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Debug|x64.Build.0 = Debug|Any CPU
+		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Debug|x86.Build.0 = Debug|Any CPU
 		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Release|x64.ActiveCfg = Release|Any CPU
+		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Release|x64.Build.0 = Release|Any CPU
+		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Release|x86.ActiveCfg = Release|Any CPU
+		{4811B6AB-D943-E9E0-ECD3-3BF20D0D6436}.Release|x86.Build.0 = Release|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Debug|x64.Build.0 = Debug|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Debug|x86.Build.0 = Debug|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Release|x64.ActiveCfg = Release|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Release|x64.Build.0 = Release|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Release|x86.ActiveCfg = Release|Any CPU
+		{6329159F-A5E2-4E45-B976-545BFE791474}.Release|x86.Build.0 = Release|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Debug|x64.Build.0 = Debug|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Debug|x86.Build.0 = Debug|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Release|x64.ActiveCfg = Release|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Release|x64.Build.0 = Release|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Release|x86.ActiveCfg = Release|Any CPU
+		{1F40F9F4-6402-4F73-982E-F5A388868651}.Release|x86.Build.0 = Release|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Debug|x64.Build.0 = Debug|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Debug|x86.Build.0 = Debug|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Release|x64.ActiveCfg = Release|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Release|x64.Build.0 = Release|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Release|x86.ActiveCfg = Release|Any CPU
+		{F4457056-1DF6-4EBB-8FC3-66D7C0BD7FBD}.Release|x86.Build.0 = Release|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Debug|x64.Build.0 = Debug|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Debug|x86.Build.0 = Debug|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Release|x64.ActiveCfg = Release|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Release|x64.Build.0 = Release|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Release|x86.ActiveCfg = Release|Any CPU
+		{CA34D222-9A1E-41EC-BD06-68E146505A6B}.Release|x86.Build.0 = Release|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Debug|x64.Build.0 = Debug|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Debug|x86.Build.0 = Debug|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Release|x64.ActiveCfg = Release|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Release|x64.Build.0 = Release|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Release|x86.ActiveCfg = Release|Any CPU
+		{35CD17A1-9387-4D8E-B70C-20583FEFDAAA}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
